+++ CHANGELOG.rst
@@ -1,5 +1,41 @@
-0.3.0 (2015-02-03)
-==================
+Changelog
+=========
 
+0.5.1 (2015-29-09)
+------------------
+* cms 3.1 compatibility fix
+
+0.5.0 (2015-19-08)
+------------------
+* added django 1.7 & 1.8 compatibility
+* fixes AttributeError with orphan plugins
+
+0.4.1 (2015-10-07)
+------------------
+* added notification config class to support custom text variables
+* allow disabling email html version
+* allow hiding of email body txt format field
+* fixed bug with serialized boolean value
+
+0.4.0 (2015-02-07)
+------------------
+* added email notification contrib app which includes new email notification form
+* added html version to admin notification email text
+* changed the users_notified field to a text field to support non user recipients
+* hides the captcha field/value from serialized data
+* cleaned up field serialization logic.
+
+0.3.3 (2015-29-05)
+------------------
+* added support for default values in selectfields, multiselectfields and radioselects (bootstrap).
+* fixed empty values in select options
+
+0.3.2 (2015-19-05)
+------------------
+* bootstrap3 support
+* added bootstrap markup templates for all field-types
+
+0.3.0 (2015-02-03)
+------------------
 * multi-boilerplate support
-  new requirement: aldryn-boilerplates (needs configuration)
+* new requirement: aldryn-boilerplates (needs configuration)